artifact: Refactor storage package structure
